requirements.txt
amqp==5.3.1
asgiref==3.8.1
attrs==24.2.0
billiard==4.2.1
celery==5.4.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
Django==5.1.1
django-cors-headers==4.6.0
django-debug-toolbar==4.4.6
djangorestframework==3.15.2
drf-nested-routers==0.94.1
drf-spectacular==0.27.2
gunicorn==23.0.0
idna==3.10
inflection==0.5.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kombu==5.4.2
loguru==0.7.2
packaging==24.1
pika==1.3.2
prompt_toolkit==3.0.48
psycopg2==2.9.9
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
PyYAML==6.0.2
redis==5.1.0
referencing==0.35.1
requests==2.32.3
rpds-py==0.20.0
six==1.17.0
sqlparse==0.5.1
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.3
vine==5.1.0
wcwidth==0.2.13
win32-setctime==1.1.0
